pyproject.toml
[project]
name = "powerpointmcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.3",
    "mcp>=1.12.2",
    "python-pptx>=1.0.2",
    "requests>=2.32.4",
]
